requirements.txt
asgiref==3.7.2
black==23.12.0
click==8.1.7
colorama==0.4.6
coverage==7.3.3
Django==5.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
flake8==6.1.0
mccabe==0.7.0
mypy-extensions==1.0.0
packaging==23.2
pathspec==0.12.1
platformdirs==4.1.0
pycodestyle==2.11.1
pyflakes==3.1.0
PyJWT==2.8.0
pytz==2023.3.post1
setuptools==68.2.2
sqlparse==0.4.4
tzdata==2023.3
wheel==0.41.2
